vendor libarchive wrapper source unzipped
The previous method of unpacking the gem and then applying patches to
fix issues is cumbersome. The libarchive gem is abandonware. It was last
updated in 2010 (version 0.1.2) and that version makes changes that
don't compile and function with this setup as it is. It seems like the
smart move here it so simply take ownership of this code for our
purposes and maintain it directly in this repository from now.